Start working with certificates
This change adds dependencies on the (now fixed) pyOpenSSL (custom branch for
now) and upstream cryptography.  This allows us to generate utf8strings in
certificates, which is a must for compatibility.

It also adds a somewhat rough certlib that contains functions to:

generate a key + req
Generate a CA cert
generate a client-only cert
generate a server-only cert
generate a client+server cert

It sets issuer, subjects and others correctly, and applies chosen constraints to
which parts of a DN are allowed.

This can later be mangled into proper setup scripts to generate CA, and to sign
requests.