Implement Subject validation based on the CA Certificate
We have a white-list of subject items that may be in the Subject,
from our CA cert we then extract the matches, and strictly validate
incoming signining requests against this.

We still care about the _order_ of fields in the Subject, and
require it to be the same order as in the CA certificate.

The actual tested function works on raw PEM data to avoid
running unit-tests on file management code.