Return json errors be consistent with validation method
Wether we should validate with "is_valid(data)" that returns
true/false or Exception chained behaviour is something I do
not have a proper opinion about.

But now we do it the same way in the same function, rather
than mixing behaviour.

The improved errors now also tell you which parts of the
Subject weren't valid, and returns it as a json error rather
than html.

This should make debugging easier.