Autosign daemon: Introduction
The autosign daemon is a multi-threaded background daemon that
connects to a caramel DB and polls it repeatedly for new certificates.

It requires all subjects to be a valid UUID, and will sign them blindly.
This is meant for running auto-registration and various services where
you first sign & accept a certificate, and then later do validation and
access control.

The autosign daemon by default generates certificates that are valid
for 3 hours.

On my development system it can sustain 222 requests/second with SQLite.
Note that if you increase the parallelism of the daemon, you risk hitting
SQLites threading issues where writes are done from different threads.