models: add an explicit query for refreshable items
Querying all items and filtering based on having certificates
later doesn't scale with amount of certificates in the DB.

This adds a custom selector that uses an SQL sub select to filter
in the selection state.

Also, this changes the default property for CSR to not-lazy load
certificates.

This is to prevent 20k new select lines when iterating over 20k
posts to refresh them.