models+tool: Improve how we list older certificates
CSR.stale(datetime) a helper to list certificates that will be
stale/expired by a certain date

CSR.most_recent_lifetime  a single-query request to get the active dates for
all certificates

Since we mix long lived and short lived certificates, it's not enough to use
`stale` in automation functions.

most_recent_lifetime would have been nice to do all the logic in the database,
however SQLite doesn't perform datetime comparisions / operations properly, so
we have to do so in code.

This is intended for an autorefresh daemon.