caramel, tools:  Make the CSR.certificates a lazy object
CSR.certificates turned out to be a major pain point once you end up with more
certificates in the database.

This turns certificates into a dynamic/lazy query. It necessitates a few
changes to when it was used as a list.

Other than this, I found a few places where we pass CSR's between thread
objects. This works "sort of" on SQLite, but fails on a multi-connection pool
against PostgreSQL.

The fix is to not pass objects between thread contexts, and only pass sha256sum
and "real" data, rather take the cost of an extra fetch of the CSR inside each
thread context inside it's own transaction.