views: Check for existing duplicates before causing SQL errors
Postgres complains more about INSERT's that break UNIQUE constraints,
and also complains in it's log file, including copies of the PEM.

This turned out to cause almost a gigabyte of log data after a benchmark
went wrong. We make a quick check to see if it's there already.

Note that it still has to handle the failure mode due to concurrency issues.