caramel-refresh: Fix continue after error
The return command when run in a subshell won't return from the function
that invoked the subshell, only the subshell itself. This change
modifies the script to not use subshells in these cases. Also, the
return status is changed to 0 so that execution of the script isn't
stopped after the function returns.