Make: Add basic Makefile with system test
The Makefile handles the setup and execution of a "complete" system test,
ensuring that caramel can be installed and run on the machine. To conduct the
test it first generate a new CA cert+key using caramel_ca with
tests/ca_test_input.txt, then a new database is initialized with
caramel_initialize_db. After the setup is complete the server is started in the
background with the generated files, the same is then done for caramel_autosign.
After that use client-example.sh to first upload a CSR, then again to check if
the server has processed our CSR, in which case we should have a file
client.csr, if not something has gone wrong. Any failiur to pass stages of the
test results in exitcode 1.

make venv-install 	setup a virtual environment(mktemp -d), install caramel here
make systest	run the system test, will install caramel if not already done
make clean		removes the virtual environment(mktemp -d)