Scripts: Move caramel_launcher.sh to scripts
Since caramel_launcher.sh is a sh script move it to the other scripts. So to
not break anything the Makefile is modified to look for the script there.
To acctually facilitate the needs of Makefile caramel_launcher takes 1 argument
which is what pserve to use, since in the testing pserve from the virtual
environment needs to be used.