Config: change CARAMEL_LOG_LEVEL from int to str
It is not very logical that the environment variable should adapt itself to the
form logging level that the argument handling use, when it can be simply a
string that describes what it is. So CARAMEL_LOG_LEVEL should be set to either
error, warning, info or debug, not sensitive to capitalisation.