Config: add dburl from arguments to bootstrap
Bugfix to be able to use autosign and tools using only the commandline. If an
argument is passed on the commandline it should trumph all other. So in
caramel.config.bootstrap we can safely overwrite the env-var CARAMEL_DB_URL
with what was passed on the commandline if there where any.

Previously:
If we want to use tools or autosign using only the commandline for
configuration, when we try to bootstrap the app it wont know what database to
connect to and crash since it can only see configuration in either an ini-file
or an env-var.